manager: use libudev directly
When using gudev we have no way of getting the kernel changes. But
we want those to see can (in the future) report thunderbolt device
activity as early as possible, to inform the user that something
is happening. This is necessary because a long time (seconds) can
pass between the thunderbolt devices are connected and they can
be used. This is due to the authorization process.