device: rework ctor to work directly with sysfs
Instead of getting the unique-id (and other properties) via udev,
we open the sysfs directory itself and then read the attribute
via a combination of openat() and fgets(). This will ensure that
the uid we obtain is in the same directory that will later be to
authorize the device (by writing to the "authorized" sysfs attr).