manager: auto-authorize devices
if the device has the policy set to DEVICE_POLICY_AUTO and
is not yet authorized and the parent is authorized, then
automatically authorize it, when it is connected.

Also when the status of a device change, and it is now
authorized, then check if any children are eligible for
auto-authorization.