device: explicitly set "stored" property
If the "store" property is changed, we also explicitly set the
"stored" property (although it is a transient property). This is
needed because the BoltDBusManagerSkeleton object created by
gdbus-bindgen will cache properties and only sent out changes if
they are changed (according to its own cache); therefore we need
to update the cache.