enum: add CONNECTING value to BoltStatus
This will indicate that an already known (stored) device has been
connected and boltd is in the processes to determine if it can
be authorized or not. This means the state flow for *stored*
devices is as follows:

a) we cannot authorize the device (e.g. auth disabled):
   disconnected (-> connecting) -> connected

b) we can authorize the device and it worked:
   disconnected -> connecting -> authorizing

c) we tried to authorize the device but it failed:
   disconnected -> connecting -> auth-error

For new devices (not stored) nothing should change.
   connected -> authorizing -> success: authorized
                               failure: auth-error

NB: This just prepares the enum value but does not implement the
aforementioned behavior yet.

In terms of bolt_status_is_connected the new CONNECTING state is
considered to indeed be connected. This might change in the future.