manager: move auth of children to …_status_changed
If the status of a device is changed, and the new status is AUTHORIZED
then we also check if we now can authorize any children of that
device. Currently this is done in the udev change event handler, i.e.
handle_udev_device_changed; Now we do that in the signal handler of
the device status change signal handle_device_status_changed. The
latter will be called anyway as a consequence of the udev event, if
that leads to a status change. The change puts reaction to device
status changes all in one place, so the code is more organized.