test: integration: check auto-authorization
Create stored devices, start the daemon and check that the devices
that can be authorized are and that the devices that can't are not,
e.g. they have no key and we are in SECURE mode or the parent of the
device can't be authorized.
Also check that we react to external authorization changes and react
properly.