device: report more specific authorization errors
Instead of always reporting BOLT_ERROR/BOLT_ERROR_FAILED we now
propagate the more specific error codes we get from the underlying
functions. This includes, NOKEY, BADKEY errors.
Additionally, in the case the kernel reports EINVAL, we try to
resolve that to a more specific error: either BADSTATE, if the device
is already authorized, or AUTHCHAIN, if the parent device is not
yet authorized.
We also don't prefix the error with a "failed to authorize device"
error anymore, that should be handled by the caller.