manager: don't set storetime, let put_device do it
Now that bolt_store_put_device will take care of correctly setting
the storetime to the current time if it is not set, we don't need
to do that in the manager.
As a positive side-effect it will also be emitted together with
the rest of the store related property changes.