manager: auto import boot-authorized devices
If we detect a new, i.e. not stored, device that is a) authorized
and b) has the boot flag set, we auto import the device.

We also check that we are not in SECURE mode and have no key. The
idea here is that even though the device was already authorized at
boot, we could still create and in the future use a key in SECURE
mode. So hopefully the user will re-connect the device when the
session is not locked and then we automatically get the Enroll
command from the shell, which in turn will generate a key for
the device and imprint it.
The check is currently theoretical because the key loading function
will return an error if the key is empty, but that might change in
the future so that check is good to have.