manager: label new devices, detect duplicates
Instead of cleaning up the name and then storing the cleaned-up
name in the database, we store the name, vendor as we get them
from sysfs. We then cleanup the name, detect duplicate name, vendor
combinations and label the device with the cleaned-up name with
an optional suffix in case of duplicates.
Adapt the integration test to reflect those changes.