boltd: introduce BoltDomain to device and manager
BoltManager will no have a representation of the thunderbolt domains
in the system. This is done by having one BoltDomain object per
thunderbolt domain udev device. The list of domains is being updated
via uevents.
When thunderbolt devices are connected the corresponding domain, in
form of the BoltDomain representation, will be passed to the specific
BoltDevice.