exported: helper to manually override properties
The normal way to override existing properties for a gobject is via
g_object_class_override_property, which internally generates a
GParamSpecOverride and call g_object_class_install_property. In bolt
we normally use g_object_class_install_properties which expects an
array of GParamSpecs. Same for bolt_exported_class_export_properties.
Therefore we craft a helper function that looks up the property and
creates the GParamSpecOverride but does not yet install it, so we
can then install it together with all the other normal properties.