exported: support object-path auto-generation
Instead of relying on a supplied object path in _export, a valid
object path will now be auto-generated as a combination of a
per-class object pass (can be set via a new class-method called
+bolt_exported_class_set_object_path) and the individual instance
object-id property. Both components are optional though.
The object-id is transformed to remove certain invalid characters.