exported: don't hold reference for dbus connection
BoltExported was holding an extra references while the object
was exported. As a result, objects that were exported were not
freed unless the owning object unregistered the object before
releasing its reference. This is cumbersome.
Now that BoltExported un-registers to object on finalize, there
is no danger that any vtable method will be called on an object
that has been finalized. We can therefore get rid of the extra
reference simplify object lifetime management.