sysfs: add functions read and write boot acls
Reads the boot_acl attribute of a udev device and parses it, returning
a GStrv; the length of the GStrv can be used to determine the maximum
number of allowed entries.
The writer functions takes a GStrv, formats and writes it to the
boot_acl attributed located inside the given device (specified via
a path).