manager: emit property notify for security-level
In the one case where we switch the security level from 'unknown'
to the correct value (i.e. when a domain is registered for the
first time) emit a proper property change notification.
This will also emit a dbus property changed signal for the
corresponding dbus property SecurityLevel. This will actually
matter only if the manager is already exported when the security
level is determined; this could be the result of force-power
not being supported by the system.