power: introduce RAII (BoltPowerGuard) object
Instead of having a simple function that turns force power on and
off we now use a combination of a resource-acquisition is
initialization object and internal reference counting. Each RAII
objects (BoltPowerGuard) represents one "reference" of a client
that wants to turn on force_power. When the object is released
the reference is removed. If there are zero references left
then force_power is shut down.