enum: add BOLT_FORCE_POWER_WAIT to BoltPowerState
Add a new entry to the BoltPowerState which indicates that all
no more guards/locks for force_power are active, but the force
power is still ON and a timeout has been scheduled for it to
be switched off. This is so that udev events that might still
come in can be processed and the controller can shut down
gracefully after all processing by clients was done.