power: delay switching force_power off
Instead of switching power off directly, when the last client
released its PowerGuard, add a mechanism to be able to delay
the the actual power down of the controller (switching
force_power to OFF) by a given amount of time.
Currently the delay is set to zero, mimicking the old no delay
situation; this is done to not make the tests fail.