power: reset the wait timeout when devices appear
When new thunderbolt devices appear, we reset the timeout, so
we make sure we have enough time buffer to process all uevents
with the controller still being powered.
This is very similar to what fwupd has done to avoid a very similar
situation in fwupd[1]. The one difference is that currently in the
code in the BoltManager, we actually wait for the domain to appear.
But then we have the problem that after the manager has relased its
power-guard, the queued "added" uevents for the domain and host and
will be processed by the manager and we want the thunderbolt
controller to be powered while this happens. So with this change
we should have 20 seconds for each device to process it.
[1] https://bugzilla.kernel.org/show_bug.cgi?id=199631#c41