power: save and recover force-power state
We record the current state of force-power by creating a marker
file ("on") in the state dir ("/run/boltd/power") that is present
iff we turned on force-power. The idea is that on crashes of boltd
we can recover the state. This all is necessary because the sysfs
file to force-power the controller is write-only.