power: implement monitoring support for guards
Power guards can now be monitored via a FIFO; this means a file
descriptor for writing is opened and the reading side is monitored.
When an event is observed on the reading side, the guard will
release the reference held for the watch. This can be used
to implement a DBus API where closing the monitoring fd will
release the power guard.