manager: add ForcePower dbus method
Add a dbus method ('ForcePower') that allows clients to force power
the thunderbolt controller. The return of the function is a file
descriptor. That fd will act as a guard/lock for the force power
state, i.e. ss long as that fd is open the force power is guaranteed
to be set to ON in boltd.