power: do not remove state if fifo is active
When we are cleaning up the guard, but the guards has an active
fifo, then we do indeed not remove the state (and thus basically)
the guard. In practice that should only happen if a client still
has the write side of the FIFO open and bolt is shutting down,
either due to --replace or due to a crash. In that case we want
to be able to recover the guard.