boltctl: initialize correct variable in error case
When the list_domains () call fails, the domains variable will be
empty, but the rest of the code relies on it being a valid array.
So we initialize it with a empty array. The code already did that
but was initializing the WRONG array (devices) instead of the
correct one (domains).