device: auto-generate dbus path based on uuid
Use the auto-generation of the object path feature that comes with
BoltExported. Instead of generating the object path locally and
then storing it twice, in BoltDevice and in BoltExported, we
specify the base path (bolt_exported_class_set_object_path) and
then add a new "object-id" property to BoltDevice that will be
used by BoltExported to generate a unique, per device object path.
The result will be identical to the old one.
The behavior of bolt_device_get_object_path is unchanged though:
only of the device is exported it will return a non-NULL path.
Therefore BoltManager does not need to be changed at all.