device: store (update) timestamps
When a device is connected, authorized or updated through udev and
as a result conntime or authtime gets set or updated we store that
new timestamps in the store (BoltStore).
Also, when a device gets disconnected, we *don't* set the conntime
and authtime to 0; this means that now conntime and authtime have
the meaning the consistent "last connected", "last authorized"
meaning.