store: multi-value version of {get, del}_times
Analogues to bolt_store_put_times the multi value version get_times
and del_times, will warn if an error occurs and no error argument
is provided, but keep going. If an error argument is provided the
whole process is aborted on the first occurrence of an error, unless
it is the not-found error.