device: upgrade to key authorization if possible
If a device gets "manually" authorized via the d-bus "Authorize"
method, the controller is in SECURE mode, but there is currently
not a key stored for the device, a new key will now be generated,
if possible. The preconditions are: the device is stored and it
allows for key authorization.