power: setup io channel earlier
This will prevent the fd to be leaked and not closed in case opening
the writer side fails. Uncovered by coverity (cid 320275).
We explicitly set the fd to -1 and add a comment, even though we
will overwrite it in the next line again; hopefully the compiler
will optimize it away but it makes it explicit what is going on.