meson: require meson version >= 0.44.0
Mainly because of "warning"; which seems a overkill to bump the
version just to have "warning" instead of "message", but a quick
check revealed that it seems widespread enough, e.g. F28 (0.47.2),
Ubuntu bionic (0.45.1), Debian unstable (0.47.2) and Arch (0.47.2).

A quick patch to revert that change, if it should be necessary is:

diff --git i/meson.build w/meson.build
index a224d05..fa15a17 100644
--- i/meson.build
+++ w/meson.build
@@ -2,7 +2,7 @@ project('bolt',
   'c',
   version: '0.5',
   license : 'LGPL-2.1+',
-  meson_version: '>= 0.44.0',
+  meson_version: '>= 0.42.0',
   default_options: ['warning_level=1',
                     'c_std=gnu99',
                     'buildtype=debugoptimized'])
@@ -406,7 +406,7 @@ if res.returncode() == 0
   endforeach
 else
   msg = '@0@'.format(res.stderr().strip())
-  warning(msg)
+  message(msg)
 endif

 # contrib