data: directory management via systemd
Prepare for systems where boltd itself can not create the toplevel
directories in /var (localstatedir) and /run (runtimedir). systemd
will take care of creating the directories for us. In case it does
not do so, boltd will anyway attempted to do so and only fail if
a) the directory does not exist and b) could not be created.