data: unit: depend on polkit.service
PolicyKit is used to authorize dbus method calls. Currently the
dependency on polkit is implicit, i.e. boltd creates a polkit
authority (via polkit_authority_get_sync) and that in turn will
use dbus activation to start the service. As a result there is a
~20 seconds gap during bolt initialization (which makes boltd
startup take ~25s).
Now the dependency is explicitly encoded, which should make it more
transparent what is actually going on. We also reduce the "blame"
time to sub-seconds (so we are not to be blamed anymore):

BEFORE:
systemd-analyze blame
           25.165s bolt.service

AFTER:
systemd-analyze blame
           191ms bolt.service