common: cleanup function for file descriptors
Sometimes, when dealing with file descriptors, we don't really
care about the return value of close () and prefer to have it
closed automatically when we leave the scope. A valid case when
opening a file read-only. For this we define "bolt_autoclose",
very much in the spirit of g_autofree.