manager: lookup domain for devices earlier
We need to domain when handling "add" device uevents for both
cases: device_added and device_attached, i.e. the former for new
and unknown devices, the latter for stored and now connected devs.
So do that in the central place and pass it as a function arg
to both functions.