manager: create domains objects on demand
Instead of creating the domain objects when we first see the
thunderbolt_domain udev device, the manager will now delay the
creation until the first actual thunderbolt_device udev device
that is attached to the domain will be created. This must be the
host device for the root switch.
This change is a preparation so that we can uniquely identify
domains via their root switch devices, which carry a unique_id,
in contrast to domains, which don't.