domain: online updates for bootacl entries
Functions to add and remove single uuids to the bootacl of a given
domain, if the domain is online. The allocation strategy, given a
bootacl strv with n entires [0, 1, ..., n-1], is the following:

  - fill empty slots from left to right
  - call an allocation hanlder, with the bootacl list, the new uuid
    and a pointer to a slot (out parameter)
  - if no allocation handler is registered do FIFO:
    shift all entries by one to the right and replace the last

The allocation handler is implemented as a signal, where the first
signal handler that returns TRUE wins.