domain: bootacl supports offline modifications
When a domain is online, the changes are written to a journal (of
type "bootacl" with the domain's uuid as the journal name). When
the domain comes back online, the journal entries will be applied
to the bootacl (which might have changed, i.e. booting a different
installation or operating system). Hence we apply the journal on
top of the newly read ACL and the result might be different from
the acl that was calculated and "cached" in the domain.