test: integration: fix arg order in Recorder
When waiting for events, any new events are match against a list
of events we are waiting for. The match is done via event_match()
which expects (event, target) as arg. This is important because
if target.detail is None the detail is ignored when doing the match.
Since the order was swapped and we supplied (target, event) the
detail was never ignored but always match.