test: check bootacl allocator
The number of slots that a thunderbolt controller can store usually
is limited (e.g. 16); this means that if all slots are already taken
and a new device ought to be added a decision needs to be taken of
which device to replace in favor of the new device. The default
mechanism is FIFO, but that might not be a good decision because
the first device added might also be the most used one. Therefore
the BoltDomain supports installing custom allocator (in form of
signal handler) that can instead overrule the FIFO decision;
check that this works as intended by installing a custom allocator
and then always returning the first (e.g. 0) slot.