test: integration: enroll with manual and auto
When testing device enrollment, the policy for the device is now
alternated between "manual" and "auto", the result of this is that
half device will get added to the bootacl and the other half not.
Additional half the device will automatically get authorized and
the other half not, so we in total cover more cases.