boltctl: don't leak formatted uuids
When formatting uuids, the bolt_uuid_format function returns a newly
allocated string. All of those are currently leaked because they are
directly passed to printf. Instead there is now a global hash table
that will store the mapping of original uuid to formatted uuid and
on program exit the hash table, together with the strings it contains,
is freed.
The was introduced as a misguided optimization :/