power: check for proper support in acquire_full
bolt_power_acquire_full now checks directly if force-power is
supported and returns immediately if not. That makes force-power
support a pre-condition for bolt_power_switch_toggle, where the
proper check is turned into pre-condition check. In wait_timeout
not having support is checked and ignored, because the warning
that force-power went away while being turned on was already
issued in the udev event handling code.
Now that there is no warning in acquire_full anymore, because most
clients of the function already warn on error when calling it, add
a todo to the dbus method hanlder to add logging there.