test: more robust permutation test
The identity permutation is also a valid random permutation, so
instead of doing the permutation only once and then checking if
an permutation actually happened do it in a loop and check that
the identity permutation only occurred up to a threshold.
Currently the repetition is rather low but to account for that
the threshold is quite high.