store: return NULL for empty bootacl entries
When the domain is stored and the bootacl is NULL, it will result
in "bootacl=" in the key file. When this later read back, it will
result in a non-zero, but empty (i.e. 0 elements) bootacl property.
As a result the BoltDomain will report that bootacl support is
present, but many of the internal functions will fail, if bootacl
has no usable slots. Therefore when the bootacl is read back and
is empty, we set it to NULL.