test: integration: auto import in secure mode
Tests the auto-import code paths but for secure mode. For now,
importing devices in secure mode should not happen because auto
import only happens in secure when the device has a valid key
is authorized and the boot attribute is set. Currently this
should never be true at the same time.