common: helper to create on GError for errno
Like g_set_error but will create a G_IO_ERROR with the error code
derived via g_io_error_from_errno. Returns TRUE if errno is 0 and
will not do anything if error is NULL.
NB: will not create a error message from the errno, but the '%m'
extension should work (and is tested for).