manager: update security when domains connect
Extract the infer security logic out into a separate function and
connect to the 'notify::security' signal so it will be called when
the domain is registered as well as when an already registered
domain updates its security level (as a result of it being connected).

This fixes a bug where the security level of the manager will would
not get updated and would stay as 'unknown' because a previously
stored domain, that is already connected when bolt starts, is not
used to update the global security level.