domain: clear error var after warning
In case we got an error reading the boot acl from sysfs when
constructing a new BoltDomain from sysfs (domain_new_from_udev)
the error variable did not get reset; do that, in order to be
able to re-use it later.
Properly include the LOG_TOPIC in the warning as well.