enums: introduce new BOLT_POLICY_IOMMU policy
A new policy for devices designed for the new iommu protection
mechanism of modern hardware/kernels. The whole point of this
new security scheme is to NOT have to ask the user for device
authorization but automatically authorize/enroll newly connected
devices. On subsequent connects these devices should automatically
be authorize as with the AUTO policy. But on the other hand, when
on future boots the iommu support is absent again (disabled in the
BIOS) the device CANNOT be trusted anymore and therefore should
NOT be automatically be authorized.
Therefore the IOMMU policy should be like AUTO when iommu is active
but like MANUAL if not.