manager: auto-authorize for iommu
When a device has the IOMMU policy set and the domain it is connected
to has iommu enabled, automatically authorize the device. NB: the
usual security level checks still apply, i.e. if the domain is in the
SECURE mode and the device as no key, NO auto-authorization will
happen.