names: expose object path generation method
Export the method used in BoltExported to generate the object path
for a object based on the 'base path' (e.g. /org/freedesktop/bolt)
and the 'object id' (e.g. the uuid of the device or domain).

For the origin and other justification of why the generation is
done in that way see the original commit that introduced that
feature in BoltExported: 298b180acfe8654be362647d8ecd447c4311a3b7