client: method to enroll a chain of devices
Add a helper function that takes a array of uuids and will enroll
them (from 0 to array->len) one by one. It is mainly meant as a
means to enroll a chain of devices, i.e. a target and its parents.

The implementation was done in a generic way so it can be shared
with a future authorize_all function.