boltclt: fix order of authorization for '--chain'
Use the new bolt_client_list_parents function that will return
parents from the immediate one to the host. Walk this parents
list in the reverse order so to start with the root and end with
the immediate one, skipping already authorized/enrolled devices.
This will fix boltctl {enroll, authorize} --chain where devices
were not authorized in the correct order.