manager: adjust policy to 'iommu' if iommu is on
When enrolling new devices with a default policy and iommu is turned
on, always adjust the policy to 'iommu'. This might lead to a
situation where a device needs to be manually authorized when it was
enrolled while iommu was on and iommu is now disabled. While not ideal
this seems to be the most secure situation to end up in.