bolt-mock: use mock rundir
Instead of using the system wide rundir, also use a mocked rundir.
Normally they should not really conflict because the boltd started
via bolt-mock will replace the currently running one, but it is
good to not interfere with the system anyway. Maybe later there
could be an option to explicitly use the system rundir.