test/integration: add workaround for flaky tests
When executing the integration test suit as a whole there seems to
be issue with umockdev not properly being cleaned-up (or set up)
and at least two tests would fail. This never occurs when the tests
are executed individually. Manually running the garbage collector
seems to reliably fix the issue.