proxy: use static strings for variant → value conv
In get_dbus_proxy, special case the string → string property
when "converting" from the GVariant to the GValue. As the comment
in the source code states, this assumes that the lifetime of the
GValue is bound by the lifetime of the GVariant. The latter could
be invalidated by a property change event for the GDBusProxy,
which could happen with main loop iterations.