common/io: add assertion for static analysis
Make the clang static analysis happy by asserting that error is
not NULL before de-referencing it. This can indeed never happen
because error is set in all cases that NULL is returned from
bolt_read_value_at.
The static analyzer thinks that, when fgets returns something
non-NULL, it might end up in a NULL return with error not set,
but indeed we do g_strdup(line), where line is a local char
array.