manager: persist the host device
When the thunderbolt controller is in non-native enumeration mode,
it is only powered by the firmware when something is plugged into
the thunderbolt port. In those cases the domain and device nodes
are missing in sysfs and we have no information about the host
device, including the (new) generation attribute. Always persisting
the host solves that issue and does no harm in native enumeration
mode either.