store: fix error handling on put for stored devs
When a device is stored, the store tries to load the keyfile for
the devices, which on the first store, will result in a not-found
error. This is the error that should always be ignore and all
other errors be reported.
Additionally, reset the error immediately instead of on-demand
later, in case some other code wants to use that error variable
in the future.