manager: add 'generation' property
Add a global 'generation' attribute ('Generation' on the D-Bus) that
will proxy the generation attribute of the host controller, or the
maximum generation, in the case that there is more than one. It is
done for convenience so that clients, interested in detecting USB4,
don't have to lookup the host device.