test/integration: new tests for generation updates
Check that updates of the generation attributes happen and are
emitted: when a host was stored without a proper generation
attribute (i.e. 0) but then gets updated (say after the kernel got
upgraded to 5.5) the generation attribute of the host but also the
global one should get updated and property change events emitted.