device: update the db entry on generation updates
When the device attributes change on connect, which currently only
is tracked for the 'generation' property, update the store entry.
A typical case is a previously unsupported kernel attribute, like
generation, that is now available due to a kernel update and thus
changes its value. This change needs to hit the store.