ci: fix permissions on build dir for travis
The user on travis is travis:travis, with uid/gid 2000:2000. Thus
the build directory gets created as "drwxr-xr-x. 1 travis travis".
Inside the container the user is root:root (0:0). Now that it is
run with --cap-drop=dac_override this will prevent root from
actually creating files in the build directory. Fix that by making
root the group owner of the build directory.