test/integration: base uuid of TbHost on TbDomain
The uuid of the TbHost now has a fixed prefix and a suffix that is
based on the domain index. Additionally the name of the TbHost can
also be supplied during object creation. Both changes are needed
to have more than one TbDomain connected, because otherwise name
or uuid would conflict.