power: refactor guards to not have power reference
Guards are meant to be used as in a RAII-like fashion, i.e. they
the resources they are "guarding" is bound to their lifetime.
Previously PowerGuard was tied to BoltPower by holding aa reference
to the former and during object destruction (dispose) it would call
a special BoltPower method (bolt_power_release) to indicate that it
was releasing its guard. Convert this to using a signal, which has
the advantage of 1) removing the tight coupling between BoltPower &
PowerGuard and 2) multiple clients can listen to and react to the
guard being released. Adapt the tests for this.