power: guards should not remove the parent dir
The idea originally was that when the last guard is removed the
parent, i.e. the state dir, gets removed as well. This currently
does not happen because the state file for BoltPower ('power/on')
is removed only *after* the guard state file is removed. Thus,
the dir is never empty. Additionally if the dir was to be cleaned
up, the next guard creation would *fail*, because the state dir
is currently only created during BoltPower initialization.