guard: don't recover guards without fifo
When recovering guards, ignore all guards that have no fifo. This
should in reality not be any different to what was already being
done, since there are only two different paths that create guards
currently:
  - internal ones, created by boltd itself during initialization
    for domain discovery
  - external ones, created by clients via the D-Bus APi, which
    all will have named pipes

Remove the test that was relying on that code path. The test
for this should be done in the BoltGuard unit tests anyway.