guard: don't clean fifo in released state
If the guard is being released but the fifo and thus the watch are
still active, the guard_watch_release function will be called due
to the removal of the watch in the finalizer. In that case the fifo
was not being cleaned up on purpose so we just skip the cleanup.