power: remove watching for the process of guards
The idea behind was that a guard could be active for a process and
when the process dies the guard needs to be removed. This could in
theory happen for a guard of an external process that is does not
have a fifo. In reality this case is impossible, because any guard
that is not created internally, was created via a D-Bus call which
in turn will always create a fifo. The fifo itself is sufficient
to also catch the death of the process even when it happens during
a bolt crash because when bolt re-starts and starts to monitor the
fifo again it will receive the corresponding close event.
Therefore remove process watching from BoltPower.