common: introduce new BoltLinkSpeed struct
Introduce a new BoltLinkSpeed struct that represents the speed of a
link between two thunderbolt/usb4 devices in terms of lanes and the
rate (speed) of the individual lane. See the previous commit for a
description of lane bonding and transfer speeds. The BoltLinkSpeed
struct is registered as a boxed type in the type system so it can
be used as a GObject property. Wire conversion functions are also
provided so it can be used in custom wire conversions and thus be
used with BoltExported.