sysfs: support reading the link speed
Add support for obtaining the link speed for a device via the
existing bolt_sysfs_info_for_device function. If the 'full'
argument is 'TRUE' the new 'linkspeed' member will contain
information about the number lanes and speed of the lanes,
if the corresponding sysfs attributes exist; otherwise they
will be set to '0'.