boltctl: show the link speed, if available
If the link speed contains non-zero values for lanes and speed
show the details of rx and tx speed. Insert before authflags to
to be sore that something is before us and something is after
us so we don't need extra logic to know what tree element to
insert ("branch" or "right").