test/integration: ability to skip store perm check
Add the ability to skip the store failure tests. This is necessary
because they rely on DAC (discretionary access control) of the
file in the store, i.e. the test removes the permission to write
to the file. If the test executor has CAP_DAC_OVERRIDE, it can
bypass that and thus the test wont work.
Consolidate the two store failure checks (label, policy) into an
extra test and skip that if the caller has CAP_DAC_OVERRIDE.
NB: "ninja -C build test" will still show the test as Ok (and the
log will tell it was skipped) because all the tests are executed
individually and the result then will be Ok(skipped=1) which is
not picked up by meson.